Extract form logic to core package
