+++ packages/core/src/types.ts
@@ -1,6 +1,21 @@
 import { Atom } from "./atom";
 import { State } from "./state";
 
+export type FilterFirstElement<T extends unknown[]> = T extends [
+  unknown,
+  ...infer R
+]
+  ? R
+  : [];
+
+export type FilterFirstTwoElements<T extends unknown[]> = T extends [
+  unknown,
+  unknown,
+  ...infer R
+]
+  ? R
+  : [];
+
 export type extractStateValue<T> = T extends Atom<infer V>
   ? extractStateValue<V>
   : T extends State<infer V>
