+++ packages/core/vite.config.ts
@@ -8,9 +8,14 @@ export default defineConfig({
     lib: {
       entry: path.resolve(__dirname, "src/index.ts"),
       name: "Iniz",
-      fileName: (format) => `index.${format}.js`,
+      formats: ["es", "cjs"],
+      fileName: (format) => `[name].${format}.js`,
     },
     rollupOptions: {
+      input: {
+        index: "./src/index.ts",
+        form: "./src/form/index.ts",
+      },
       // make sure to externalize deps that shouldn't be bundled
       // into your library
       external: [],
@@ -18,6 +23,9 @@ export default defineConfig({
         // Provide global variables to use in the UMD build
         // for externalized deps
         globals: {},
+        // By default chunk's filename will be the file will most code without format, causing different formats to overwrite each other.
+        chunkFileNames: `chunk.[hash].[format].js`,
+        inlineDynamicImports: false,
       },
     },
   },
