+++ packages/react/src/form/index.ts
@@ -1,6 +1,2 @@
-export * from "./array";
+export * from "@iniz/core/form";
 export * from "./components/Field";
-export * from "./field";
-export * from "./form";
-export * from "./group";
-export * from "./validators";
