+++ packages/react/src/form/types.ts
@@ -1,20 +1,5 @@
 import { Atom, State } from "@iniz/core";
 
-export type FilterFirstElement<T extends unknown[]> = T extends [
-  unknown,
-  ...infer R
-]
-  ? R
-  : [];
-
-export type FilterFirstTwoElements<T extends unknown[]> = T extends [
-  unknown,
-  unknown,
-  ...infer R
-]
-  ? R
-  : [];
-
 export type extractStateValue<T> = T extends Atom<infer V>
   ? extractStateValue<V>
   : T extends State<infer V>
